Update design for support view and organize files
