+++ Projects/Home/Sources/Screens/HelpCenter/ReusableComponents/HelpViewSource.swift
@@ -1,0 +1,20 @@
+import SwiftUI
+import hCore
+import hCoreUI
+
+enum HelpViewSource {
+    case homeView
+    case topicView
+    case questionView
+
+    var title: String {
+        switch self {
+        case .homeView:
+            return "home view"
+        case .topicView:
+            return "topic view"
+        case .questionView:
+            return "question view"
+        }
+    }
+}
